refactor(logger): separate from core
